socketio_demo/new_requirements.txt
altair==5.5.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dnspython==2.7.0
eventlet==0.39.0
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.25.1
networkx==3.4.2
numpy==2.2.2
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playsound==1.3.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
silero-vad==5.1.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
soundfile==0.13.1
streamlit==1.42.0
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
